lab2 syscall

+++ kernel/proc.h
@@ -103,5 +103,5 @@ struct proc {
   struct file *ofile[NOFILE];  // Open files
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
-  uint8 traceMask;               // system trace systemcal id
+  int traceMask;               // system trace systemcal id
 };
